requirements.txt
Django==1.11.14
django-cors-headers==2.1.0
django-filter==1.0.4
django-extensions==1.7.9
djangorestframework==3.6.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
python-dateutil==2.7.3